feat: time_limit 소수점 허용 (#43)
## 추가된 기능
time_limit 인자에 float을 허용합니다.

## 테스트 방법
서버 띄운 후 bru run 또는 time_limit 에 실수를 넣어서 직접 요청

+++ hodu-core/src/lib.rs
@@ -24,7 +24,7 @@ pub struct MarkParams<'a> {
     pub code: &'a str,
     pub stdin: &'a str,
     pub expected_stdout: &'a str,
-    pub time_limit: u32,
+    pub time_limit: f64,
     pub memory_limit: u32,
 }
